Alquiler y Devolución
El panel alquiler(ahora PanelAlquiler2) esta conformado por cuatro
paneles separados.  Codifique la devolución pero todavía no genera
registro en la base de datos.
Agregue la clase Multa y BaseMulta para poder completar la devolución(ya
calcula que hay una multa pero no genera registro).

+++ src/javaapplication39/Bases/BaseAlquiler.java
@@ -100,7 +100,7 @@ public class BaseAlquiler {
                 nuevo= new Alquiler();
                 nuevo.setCodClie(set.getInt("codclie"));
                 nuevo.setFechaAlquiler(set.getString("fecha_alq"));
-                nuevo.setMonto(set.getInt("monto"));
+                nuevo.setMonto(set.getFloat("monto"));
                 nuevo.setFechaEntrega(set.getString("fecha_ent"));
                 nuevo.setCodAlq(set.getInt("codalq"));
             }
@@ -127,7 +127,7 @@ public class BaseAlquiler {
                 nuevo.setCodClie(cod);
                 nuevo.setFechaAlquiler(set.getString("fecha_alq"));
                 nuevo.setFechaEntrega(set.getString("fecha_ent"));
-                nuevo.setMonto(set.getInt("monto"));
+                nuevo.setMonto(set.getFloat("monto"));
                 lista.add(nuevo);
             }
             set.close();
